requirements.txt
Django==3.2.18
djangorestframework==3.14
psycopg2-binary==2.8.6
aiohttp
requests
drf-yasg
channels
celery
redis